fix: removendo duplicidade de codigo
